Isolate everything related to text normalization in one place. And implement stuff for symbols
